Refactoring #1
Updated:
• Makefile
• Source files
• Documentation
• Headers (a bit)
• Tests & examples

Removed:
• MakefileTest
• MakefileExample

+++ src/lil_is_null.c
@@ -1,9 +1,12 @@
+#include <stddef.h>
 #include "../include/longintlib.h"
 
 int lil_is_null(lil_t *src) {
     // return 1 if source is empty, 0 otherwise
-    for (int i = 0; i < src->size; i++) {
+    
+    for (size_t i = 0; i < src->size; i++) {
         if (src->val[i] != 0) return 0;
     }
+    
     return 1;
 }
